// src/app/page/master/map-fakultas-program-jurusan/map-fakultas-program-jurusan.interface.ts
import {EventEmitter} from '@angular/core';

export interface mapFakultasProgramJurusan {
	kdFakultas?: number;
	kdProgram?: string;
	kdJurusan?: string;

	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
	value?: any;
}